+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Do not enlarge small attached images.
 - Replaced "Subscription expired {date}" with "Subscription expired on {date}".
 - Don't display prompt in post editor when editing existing post.
+- Show empty post editor if markdown source is not available.
 
 ## [4.7.0] - 2025-08-01